v0.6.8: match records between fasta and GFF with NCBI style headers
NCBI PGAP can produce fastas with NCBI headers (`lcl|1`) and GFFs
with plain IDs (`1`). This avoids creating duplicate records by
matching these and using a consistent contig ID when combining sequences
and GFF files.

The contig ID used in the final output SeqRecords is the first one
encountered. So if you use an external FASTA it's the IDs in the FASTA
and if you use a GFF embedded FASTA it's the IDs in the GFF.