distribute_setup.py: migrate away from distutils.
Following [PEP 632], the distutils module is deprecated.  In gff, this
module is still in use for it's logging facility, so a contemporary
approach would be to use the "logging" module instead of distutils.
logging.warning behaves the same way as distutils.log.warn; there was
also a logging.warn function but it is deprecated in favor of warning,
thus this patch does not attempt to do something smart with things
allowed by the import semantics.

[PEP 632]: https://peps.python.org/pep-0632/

Signed-off-by: Étienne Mollier <emollier@debian.org>